Updated tests to use long instead of double for task durations.

+++ swapp/src/test/java/be/kuleuven/cs/swop/ButtonMashingUI.java
@@ -110,7 +110,7 @@ public class ButtonMashingUI implements UserInterface {
         if (random.nextBoolean()) return null;
         String description = randomString();
         boolean negative1 = random.nextBoolean();
-        double duration = negative1 ? random.nextDouble() * 50.0 : random.nextDouble() * -50.0; // arbitrary
+        long duration = negative1 ? random.nextLong() * 50 : random.nextLong() * -50;           // arbitrary
                                                                                                 // constant
         boolean negative2 = random.nextBoolean();
         double deviation = negative2 ? random.nextDouble() * 2.0 : random.nextDouble() * -2.0; // arbitrary
